requirements.txt
pandas
langchain
openai
regex
numpy
python-docx